feat: add nvidia

+++ src/nvidia-cuda/install.sh
@@ -1,6 +1,8 @@
 #!/usr/bin/env bash
+#
 
 set -e
+[ "$(uname -m)" = "amd64" ] || exit 0
 
 # Clean up
 rm -rf /var/lib/apt/lists/*
